refactor(webview): extract ChatHandler from webview-provider
